+++ functions/execute.js
@@ -1,0 +1,29 @@
+const { spawn } = require("child_process");
+
+/**
+ * Execute shell command
+ * @param {String} cmd - root command
+ * @param {String[]} args - args to be passed along with
+ *
+ * @returns {Promise<void>}
+ */
+
+const exec = (cmd, args = []) =>
+  new Promise((resolve, reject) => {
+    const app = spawn(cmd, args, { stdio: "pipe" });
+    let stdout = "";
+    app.stdout.on("data", (data) => {
+      stdout = data;
+    });
+    app.on("close", (code) => {
+      if (code !== 0 && !stdout.includes("nothing to commit")) {
+        err = new Error(`Invalid status code: ${code}`);
+        err.code = code;
+        return reject(err);
+      }
+      return resolve(code);
+    });
+    app.on("error", reject);
+  });
+
+module.exports = exec;
